+++ packages/go-kosu/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## master
 
+-   Add kosulite proxy
+-   [bug] Fix panic on kosu-cli
+-   Add /store prefix to Query Path
 -   Require power=1 in genesis file
 -   [bug][\#304] Fix BigInt JSON Serialization
 -   Add InitialPosters support